Move static functions to blinkpy/helpers/util.py
